rag-backend/requirements.txt
flask
flask-cors
python-dotenv
openai
pinecone-client
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-pinecone
tiktoken
